con rutas
