+++ public/index.html
@@ -20,11 +20,12 @@
 
       Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
       work correctly both with client-side routing and a non-root public URL.
-      Learn how to configure a non-root public URL by running `npm run build`.
+      Learn how to configure <script src="js/plugins.js"></script> a non-root public URL by running `npm run build`.
     -->
   <title>React App</title>
   <link rel="stylesheet" href="css/plugins.css">
   <link rel="stylesheet" href="css/style.css">
+
   <script src="js/jquery.js"></script>
   <script src="js/plugins.js"></script>
   <script src="js/plugins2.js"></script>
