+++ src/App.js
@@ -1,15 +1,15 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import "./App.css";
-import Blog from "./Blog";
+import Blog from "./views/Blog";
 
-import ListadoPeliculas from "./ListadoPeliculas";
+import ListadoPeliculas from "./views/ListadoPeliculas";
 
 function App() {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<ListadoPeliculas />} />
         <Route path="/blog" element={<Blog />} />
+        <Route exact path="/" element={<ListadoPeliculas />} />
       </Routes>
       {/* Switch fue reemplazado por Routes */}
     </BrowserRouter>
