+++ src/views/Pelicula.js
